[SYCL] Fix handling of host-side memory in 2D memops (#8359)
The current implementation of 2D USM memory operations will launch a
kernel as a fallback mechanism. However, these interfaces are also
intended to accept pointers to USM and non-USM host memory. To allow
this, the implementation now picks different fallback mechanisms
depending on the type of pointer passed.

For fill/memset this means:
1. Use a host-task with a fill per row if the pointer is to host memory.
 2. Use backend-defined interface if available.
 3. Use auxiliary fill kernel if pointer is to device memory.

For memcpy/copy this means:
1. Use a host-task to copy memory when both pointers are to host memory.
 2. Use backend-defined interface if available.
3. Enqueue a USM memcpy per row if one pointer is to in host memory and
the other pointer is in device memory.
 4. Use auxiliary copy kernel if both pointers are to device memory.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>