[SYCL-MLIR]: Relaxed aliasing support (#8511)
This PR adds support for the FE option -relaxed-aliasing to cgeist. The
option is used to instruct SYCL-MLIR aliasing analysis whether the input
program respect the strict aliasing rules (i.e. type-based aliasing
rules).

Options processing in the cgeist driver is streamlined to allow passing
options to various parts of the optimization pipeline.

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>