[SYCL] Don't define SYCL_EXTERNAL if -fno-sycl-rdc is passed (#8537)
Please review this commit by commit because one of the commits affects a
ton of files. Thanks!

The SYCL [spec
](https://registry.khronos.org/SYCL/specs/sycl-2020/html/sycl-2020.html#subsec:syclexternal)allows
us to declare we don't support device function calls outside of the
current translation unit by not defining the `SYCL_EXTERNAL` macro.
With the `-fno-sycl-rdc` option, this is exactly that case: we do not
support calls outside of the TU, so don't define the macro.

However, we define and use the `SYCL_EXTERNAL` in the SYCL headers also,
so instead add `__DPCPP_SYCL_EXTERNAL` and use that. It is defined to
`SYCL_EXTERNAL` if `SYCL_EXTERNAL` is defined, and otherwise
`__attribute__((sycl_device))` for device code or empty for host. This
is fine because none of the uses here actually end up as cross-TU calls,
it's either linked in from the device libraries or replaced by the
SPIR-V translator to some SPIR-V intrinsic.

I added a test in `sycl/test/basic_tests/macros_no_rdc.cpp` to make sure
nobody uses `SYCL_EXTERNAL` directly in the SYCL headers in the future.

(UPDATE: This part has been removed) However, with the macro being
undefined, if a user tries to use it, they just get the classic `unknown
type name 'SYCL_EXTERNAL'` which is not super helpful. I added some
really gross code to give a better error message, but I'd like to hear
from the CFE folks if there's a better way to do it, and if not, if it's
so gross we should just drop it and rely on the normal error.


Relevant discussion: https://github.com/intel/llvm/discussions/8479

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>