[MLIR] Allow llvm.resume with non-landingpad input
This changes verification of llvm.resume to just fail if there is no
llvm.landingpad instruction in the same function with the same type as
this instruction's input, more in line with LLVM's verificatino. From
LLVM documentation (https://llvm.org/docs/LangRef.html#i-resume):

The ‘resume’ instruction requires one argument, which must have the
same type as the result of any ‘landingpad’ instruction in the same
function.

Signed-off-by: Victor Perez <victor.perez@codeplay.com>