[SYCL] Restore original CXX flags after unified runtime is built (#8606)
We have to override CXX flags to build unified runtime because that
build generates warnings and they turn into errors with werror.

But we have to restore original CXX flags after doing so. This is
expected to build post-commit: it looks like currently we use gcc to
build external projects like unified runtime and l0 loader. At the same
time clang is used to build sycl runtime. Clang doesn't understand
-Wno-stringop-truncation option which leads to an error in post commit.