[SYCL][L0] Disable native 2D USM memcpy by default
Due to a bug in memcpy 2D to and from non-USM pointers in the L0 native
region memcpy make L0 report support for this operations based on a new
SYCL_PI_LEVEL_ZERO_USE_NATIVE_USM_MEMCPY2D which defaults to disabled.
Once the bug is fixed, this new environment variable should either be
removed or enabled by default.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>