[SYCL] Make default project directory match default executable name (#8532)
Currently, the default output directory is "<file_name>.prj" when no
"-o" flag is specified. This makes it annoying for the user to run a
design compiled for simulation since the "a.out" executable will look
for a "a.prj" directory which would not exist. By making the default
project named "a.prj", we solve this issue.