[SYCL][Reduction] Implement identityless reductions (#8562)
This commit implements SYCL 2020 identities without a known or specified
identity. This is done by adding information about when identities are
specified explicitly to the reduction objects and is propagated to
reducers through the "identity container" object, which may or may not
contain an identity. Based on this information, the reduction kernels
may pick different patch depending on whether or not the identity is
available.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>