[SYCL-MLIR] Lower to pointers with correct address spaces
Before this change, we were assuming some address spaces when lowering
`sycl` operations to `llvm`.

Now, when getting a pointer to a member of a struct pointed by a
pointer in address space `n`, we return a pointer in such address
space unless the optional `targetAddressSpace` argument in passed. In
that case, we cast to the given address space.

When offsetting an accessor, we obtain the address space from a
mapping from target mode to address space, as defined in the SYCL
standard:
https://registry.khronos.org/SYCL/specs/sycl-2020/html/sycl-2020.html#sec:opencl:native-interop-kernel. Note
that DPC++ representation of the `device`/`global` and `constant`
address space is the same (1).

Signed-off-by: Victor Perez <victor.perez@codeplay.com>