[SYCL][Reduction] Fix identityless reductions with unwritten reducers
This commit changes reducers to use a new object for holding results,
namely ReducerElement. This new object either holds a value directly or
holds an optional value. The latter is used for identityless reductions
to allow cases where the user functors did not write a value into the
given reducer and as such the reducer's value should be discounted.

This new element is used inside reductions that support identityless to
correctly propagate potentially missing values throughout.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>