[SYCL-MLIR] Add missing dialects to tools and drop deprecated option
`polygeist-opt` was missing some dialects (`vector` and `spirv`),
and passes (`sycl` dialect passes). Also, some of its passes were
missing some dialects in their `dependentDialects` definitions. The
deprecated  `preloadDialectInContext` option was hiding this last
problem, so we needed to fix before dropping.

`sycl-mlir-opt` was not using this option (just passing as `false`),
so it could be removed safely.

As `polygeist` is defining additional canonicalization patterns for
operations outside the dialect, dropping this option lead to some test
failures (updated) when not loaded. This patterns should be part of a
pattern/upstreamed to their respective dialects.

Signed-off-by: Victor Perez <victor.perez@codeplay.com>