[SYCL-MLIR] Extend and fix SYCL inline pass options
We add an `-inline-sycl-method-ops` option both to the pass itself and
to `cgeist` to decide whether to inline `SYCLMethodOps` (off by
default).

The pipeline needs further modification as `-no-mangled-function-name`
is incompatible with inlining `SYCLMethodOps`, so we have to convert
to `sycl.call` beforehand.

Now the default `-remove-dead-callees` will be used if no argument is
provided.

Signed-off-by: Victor Perez <victor.perez@codeplay.com>