[SYCL-MLIR] Always run `alwaysinline` pass
Run this pass even at `-O0`, mimicking `clang`.

We add an `-inline-sycl-method-ops` option both to the pass itself and
to `cgeist` to decide whether to inline `SYCLMethodOps` (off by
default).

The pipeline needs further modification as `-no-mangled-function-name`
is incompatible with inlining `SYCLMethodOps`, so we have to convert
to `sycl.call` beforehand.

Signed-off-by: Victor Perez <victor.perez@codeplay.com>