[SYCL] Fix weak_object and owner_less for device objects
This commit fixes an issue where weak_object and in turn owner_less
would fail to construct for SYCL object types that are usable on device.
This was due to these object types only having an impl on host. As a
result the kernel compilation would fail to get the impl of these. Since
weak_object isn't intended for use inside kernels, this was fixed by
using a dummy weak_ptr during kernel compilation to make it act like
it is containing a weak pointer.

Previously this was not found in unittests because they are not compiled
with the SYCL compiler. To avoid this breaking in the future, these
tests are moved to the test-suite.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>