[SYCL] Filter out unneeded device images with lower state than requested (#8523)
When fetching device images compatible with non-input states, we can
ignore an image if another one with a higher state is available for all
the possible kernel-device pairs. This patch adds the logic for
filtering out such unnecessary images so that we can avoid JIT
compilation if both AOT and SPIRV images are present.