[SYCL-MLIR] Add missing dialects to tools and drop deprecated option (#8692)
`polygeist-opt` was missing some dialects (`vector` and `spirv`), and
passes (`sycl` dialect passes). Also, some of its passes were missing
some dialects in their `dependentDialects` definitions. [The deprecated
`preloadDialectInContext`
option](https://discourse.llvm.org/t/psa-preloaddialectincontext-has-been-deprecated-for-1y-and-will-be-removed/68992)
was hiding this last problem, so we needed to fix before dropping.

`sycl-mlir-opt` was not using this option (just passing as `false`), so
it could be removed safely.

As `polygeist` is defining additional canonicalization patterns for
operations outside the dialect, dropping this option lead to some test
failures (updated) when the dialect is not loaded. This patterns should
be part of a pattern/upstreamed to their respective dialects.

---------

Signed-off-by: Victor Perez <victor.perez@codeplay.com>