[Driver][SYCL] Allow for undefined symbols when doing host deps generation
When processing archives when performing offload, the driver will perform
a host link which generates a file that is used to determine any offload
dependencies.  This additional link step is just for gathering the dependency
information and shouldn't require the full ability to link.

Under some circumstances, for example generating an early or final image
archive for FPGA, the symbol resolution is not important.  Add a general
case of -z undefs during this internal link step and allow the final host
link to determine any unresolved symbols at that time.