Add complex group algorithms (#7120)
Extends the implementation of group algorithms by adding support for
complex types. Operations that can be done element-wise are implemented
in headers and multiplication is implemented in CUDA PI. This involved
changing the macros in CUDA PI to accommodate complex and defining
complex types to use in PI interface. The scratch space for moving data
between workgroups in the PI also had to be doubled.

This patch depends on two fixes to the NVPTX backend which are exposed
by the new code generation.

Tests are in
[intel/llvm-test-suitegit](https://github.com/ldrumm/llvm-test-suite/tree/complex_group_algo)

Co-authored-by: Tadej Ciglarič <tadej.ciglaric@codeplay.com>