[SYCL] Remove OwnZeMemHandle from USMAllocator (#7853)
OwnZeMemHandle is used to track whether an allocation should be freed by
the runtime or not.

There is no point in passing this flag to USMAllocator and then to
USMFreeHelper since the allocation cannot come from the USMAllocator (it
can only be from zeMemAlloc*).

I've thought about moving the memory allocation code to UR first, but
there are a lot of dependencies: we would have to move the
implementation of context, device, event, and queue at least. It seemed
that doing this small change first made more sense.