[SYCL] Updated aspects list to be SYCL 2020 compliant (#8751)
This patch updates the device aspect implementations to account for the
changes made in [Section
4.6.4.3](https://registry.khronos.org/SYCL/specs/sycl-2020/html/sycl-2020.html#sec:device-aspects)
of the specification. Most changed aspects have been addresses already,
so this patch specifically:

- Adds default implementations for the `emulated` and `host_debuggable`
aspects, returning false for now. These require further device- and/or
driver-specific work so are left as is for now. This fixes
intel/llvm#8324
- Deprecates `usm_restricted_shared_allocations` aspect as the concept
no longer exists, and `host` aspect as "host" device concept removed
from SYCL 2020. Deprecated aspects will be removed in later releases.