[SYCL][Matrix] Add initial get_coord API (#7851)
This patch adds an initial API for the retrieval of coordinates from a
work item element. A `get_coord()` method is added to the intel
namespace to work on `wi_element` class. Also, a relevant SPIRV op is
added, which the get_coord() gets lowered to.

This is recreated PR from my forked repo. The discussions are in the
original (closed) PR https://github.com/intel/llvm/pull/7037