[SYCL] Make internal device_global pointer mutable (#8780)
The implementation of device_global allows it to be const. However, for
cases where the device_global is shared between programs, the value of
the device_global is defined through a USM pointer inside the
device_global. Since this pointer is seen as const when the
device_global is const-qualified, the compiler may chose to optimize it
out, potentially causing invalid accesses on device. To avoid this, this
commit marks the underlying pointer member as mutable which prevents
this potential optimization.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>