[SYCL] Do not track allocation if we don't own the handle (#8792)
If we don't own level zero handle of the pi_mem then we can't control
deallocation of that memory so there is no point of keeping track of the
memory allocation for deferred memory release. More over currently level
zero context is leaked because we increase its ref count at the point
where we start tracking memory but don't decrement that ref count in
USMFreeHelper. This PR fixes that problem.