[SYCL][E2E] Use configure_lit_site_cfg in in-tree builds
This provides the added benefit that it will automatically map the
config paths from the build directories to the corresponding source
ones in `llvm-lit`. With this, one should be able to run lit tests
directly from source paths.

In standalone builds, we don't have access to LLVM's wider
configuration, so the old configuration must remain.

Note also that this changes the configured cfg file's name from
`lit.site.cfg` to `lit.site.cfg.py`, but this aligns us with the
predominant style across the llvm project.