[SYCL][Doc] Update SYCL_INTEL_data_flow_pipes extension for FPGA host… (#8789)
A continuation of #5838. Accompanying runtime change is #7468.

Add a memory order parameter to device-side read/write members and
default to sycl::memory_order::seq_cst. This parameter is in place but
not being used at this moment, it's intended for the future work.

Add host pipe read/write members with additional sycl::queue parameter.

---------

Co-authored-by: Steffen Larsen <steffen.larsen@intel.com>