[SYCL] Host pipe runtime implementation (#7468)
Disclaimer:

This work is a continuation of a previous approved Sherry's PR:
https://github.com/intel/llvm/pull/5766 and her draft work
https://github.com/intel/llvm/pull/5894.

We are implementing Hostpipes based on the Spec
[here](https://github.com/rho180/llvm/blob/rho180-pipe-design/sycl/doc/design/HostPipes.md).
OpenCL spec is
[here](https://github.com/intel-sandbox/ip-authoring-specs/blob/MJ_ChangeDocs4/Pipe/Spec/cl_intel_host_pipe_symbol.asciidoc).

The following is the outline of the design.

1. The host pipe properties need to be added to the device image
(probably similar to the previous change here
https://github.com/intel/llvm/commit/a9ad3af71cbdd5ebc7b7efeb35761aa8ea1a49d7)

2.The frontend calls the [registration of device image, maybe similar to
this
code](https://github.com/intel/llvm/blob/af858c74e4be95b2163ce6ba545ce588ff0aca4c/sycl/source/detail/program_manager/program_manager.cpp#L2014):
which is where the host pipe information is available. This is where the
mapping from host pipe name to host pipe pointer is extracted.

3. The frontend also calls `host_pipe_map::add` to register/initialize
the host pipes. This is the new function. We are not sure about the
ordering of registration of device image / registration of host pipe.
for which ever one that comes later, it need to initialize the remaining
attribute of host pipe (such as its properties).

4. The opencl runtime will need to get a cl_program object, which is
typically not available until the first kernel launch. To get a program
object early on, the host pipe name/pointer to device image mapping is
cached during registration. And when the specific host pipe is needed,
build the program and get its ocl runtime representation. This is done
in the first couple commits.

5. Since a host pipe read/write need to depend on other write operation
finishing before it (including the inter kernel write). This means the
pipe needs to know the dependency of kernel execution. For this reason,
the host pipe read and write ocl function cannot be called with no dep
event. therefore, it is implemented with handler , which is aware of the
event that it is supposed to wait upon. This is done in the "Register
new command group .." commit.

6. Unit test
    - mock a fake device image.
    - register the fake device image
    - register fake pipe with some name you specified
- fake the opencl functionality, this can be done with
unittest::PiMock::redefine

---------

Co-authored-by: Sherry Yuan <sherry.yuan@intel.com>
Co-authored-by: Ho, Robert <robert.ho@intel.com>