[SYCL] Fix PI event leak in memcpy2d device-host fallback
The fallback mechanism for memcpy2d and copy2d between host and device
enqueues a number of memcpy operations to/from device and inserts a
marker event to signify the convergence. Currently the intermediate
events are not correctly released, causing them to leak. This commit
fixes this by utilizing an existing RAII PI event managed object to
manage the lifetime of these events.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>