[SYCL] Set the bug report URL appropriately for dpc++
When llvm tools crash they inform the user to report bugs to the
cmake-configurable ${BUG_REPORT_URL}, which defaults to upstream llvm's
github issue tracker.

For dpc++ this should be the intel llvm upstream.

I wasn't sure if editing the default in the CMakeLists.txt itself was
too invasive, but could be convinced; for now the buildbot configure
script is the place.