[SYCL-MLIR] Modify MLIR optimize pipeline (#8882)
This PR attempts to change the optimization pipeline for the SYCL-MLIR
compiler. The motivation is to remove a source workaround we had to make
in SYCL-Bench in order to get the reduction performance opportunity.
The pipeline ordering between key transformation is as follow:
```
`ArgumentPromotion` -> `CanonicalizeForPass` -> `LICM` -> `DetectReduction` -> `InlinePass`
```
Rationale: 
- ArgumentPromotion to run first in order to peel struct argument into
members and add the noalias attribute to the peeled argument
- CanonicalizeForPass takes advantage of the `noalias` attribute to
probe that load of loop upper bounds using a peeled argument is
invariant. This allow WhileLICM to hoist the load and transform
`scf.while` to `scf.for`
- LICM operates on scf.for loops (or affine.for loops) and hoists
invariant operation out of the reduction loops in the bmks
- DetectReduction operates on reduction loops where invariant operation
have been hoisted and will detect array reductions and scalarize them.
- InlinePass need to run after the previous operations to avoid inlining
operations and in doing so obfuscating their sematics

---------

Signed-off-by: Tsang, Whitney <whitney.tsang@intel.com>
Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>
Co-authored-by: Tiotto, Ettore <ettore.tiotto@intel.com>