[SYCL-MLIR] Allow argument promotion to perform on linkonce_odr functions (#8893)
According to https://llvm.org/docs/LangRef.html#linkage-types:
> linkonce
Globals with “linkonce” linkage are merged with other globals of the
same name when linkage occurs. This can be used to implement some forms
of inline functions, templates, or other code which must be generated in
each translation unit that uses it, but where the body may be overridden
with a more definitive definition later. Unreferenced linkonce globals
are allowed to be discarded. Note that linkonce linkage does not
actually allow the optimizer to inline the body of this function into
callers because it doesn’t know if this definition of the function is
the definitive definition within the program or whether it will be
overridden by a stronger definition. To enable inlining and other
optimizations, use “linkonce_odr” linkage.
linkonce_odr
Some languages allow differing globals to be merged, such as two
functions with different semantics. Other languages, such as C++, ensure
that only equivalent globals are ever merged (the “one definition rule”
— “ODR”). Such languages can use the linkonce_odr and weak_odr linkage
types to indicate that the global will only be merged with equivalent
globals. These linkage types are otherwise the same as their non-odr
versions.

As argument promotion modifies all references in a translation unit, so
it is safe to operate on functions with `linkonce_odr` linkage. To avoid
conflicting with globals defined in other translation units, we modify
the linkage to `internal` after the transformation.

---------

Signed-off-by: Tsang, Whitney <whitney.tsang@intel.com>