[SYCL][E2E] Detect supported aspects at config time
This commit introduces a framework to make better use of LIT's
UNSUPPORTED test category by detecting supported aspects at LIT
configuration time. Until now the E2E tests would exit early and PASS,
which is arguably a misleading classification.

The method determines aspects at configuration time by compiling and
then running a program. The output of this program is a
semicolon-separated list of supported aspects *on the default device* as
either '1' or '0', in an order determined by the python script. The
script captures this output and sets features as available where the
corresponding aspect is '1'.

As a proof of concept, several tests which silently pass when fp16 and
fp64 are unsupported have been converted to use this system.

The method is easily extendible with new aspects. Several E2E tests make
use of USM shared allocations without first checking the device, so a
follow-up commit would likely make that aspect a feature too.