[SYCL-MLIR][LICM] Support hoisting operations that require versioning more than one accessor pairs (#8964)
Some operations may require checking overlap of more than one accessor
pairs to be hoisted. For example,
```
for(size_t i = 0; i < 8; i++) {
   A[0] = 1;
   B[i] = 2;
   C[i] = 3;
}
```
To hoist `A[0] = 1;` out of the loop, LICM need to ensure that `A` and
`B` don't overlap, and `A` and `C` don't overlap.

By default, LICM only hoist operation that require checking of one
accessor pair. It can be changed by `-licm-sycl-accessor-pairs-limit`.

---------

Signed-off-by: Tsang, Whitney <whitney.tsang@intel.com>