[SYC] [L0] Enhance interop queue interface to choose standard or immediate commandlists (#7196)
This change enhances the interop make_queue interface to select whether
the queue should use standard or immediate commandlists. An additional
boolean parameter chooses immediate commandlists if true. The change
also modifies the L0 plugin so that use of standard or immediate
commandlists is selectable on a per-queue basis, instead of globally.