[SYCL-MLIR] Add `sycl.nd_item.get_group_linear_id` LLVM conversion
This conversion pattern reuses `sycl.group.get_grou_linear_id`'s
pattern. In order to do that, a `builtin.unrealized_conversion_cast`
is inserted to perform the conversion from the converted type to the
sycl one. This builtin will become a dead value as soon as this pass
ends, which allows a legal and easy removal, so it is not included in
the tests.

Signed-off-by: Victor Perez <victor.perez@codeplay.com>