[SYCL] Throw for info::kernel::num_args
According to SYCL2020 spec:
This descriptor may only be used to query a kernel that resides in a kernel
bundle that was constructed using a backend specific interoperability
function or to query a device built-in kernel, and the semantics of this
descriptor are defined by each SYCL backend specification. Attempting to use
this descriptor for other kernels throws an exception with the errc::invalid
error code.

Fow now the only one way to create such kernel bundle - use
make_kernel(). And creating bundle from built-in kernels is not fully
supported, so it should throw in most cases at the present time.