[SYCL] Set compiled program for hostpipe
HostPipe implementation should pass the compiled program to backend
when invoke the read / write operation.

There are two scenarios for HostPipe read / write:
   * Invoke device kernel before HostPipe read / write
   * Invoke HostPipe read / write before device kernel.

For the 1st scenario, the program should be compiled before HostPipe
read / write. We record the program in hostpipe entry.

For the 2nd scenario, the program should NOT be compiled. We need to
compile the program first and record the program in hostpipe entry.
The compiled program will be cached for future device kernel.