[SYCL-MLIR] Drop SYCL3DGridOps returning an index (#9043)
SYCL3DGridOps come in two flavors:

- One returning an `index` and receiving an `i32` argument (the
component of the 3D grid element to query);
- One returning an `sycl.(id|range)<(1|2|3)>`, the 3D grid element
itself.

The first flavor is dropped in this commit as:

- It does not represent any SYCL construct: whereas the second
represents how the `id`/`item`/`nd_item` input to a kernel is built, the
one returning an `index` does not represent any SYCL construct.
- It cannot be lowered to LLVM: Due to the OpenCL-SYCL different unit
strides conventions, we would need to perform a mirroring of the
components when lowering. This mirroring depends on the number of
dimensions of the kernel. As we do not have that information attached to
the operation, we cannot perform the mirroring, so the lowering cannot
occur.

This commit also drops the traits used to verify the type of the
operations, as this is no longer needed due to the introduced simplicity
of the operations.

Signed-off-by: Victor Perez <victor.perez@codeplay.com>