[SYCL][Reduction] Avoid implicit atomic64 requirements
Some of the existing reduction strategies use atomic operations on
partial reduction results. However, for reductions on 64-bit values this
implicitly adds a requirement that the corresponding device supports
aspect::atomic64. This commit adds additional logic to select different
strategies based on this.
Note; one of these naively chooses another strategy if the type is
64-bit without checking the support at runtime. Follow-up patches should
refactor the strategy selection, allowing this to do appropriate
runtime checks.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>