Revert "[SYCL] Represent JointMatrixINTEL type as extension type (#8343)"
It appears to be, that mem2reg and SROA passes can't handle
target extension type properly. It means, that with turned on
optimizations alloca/load/store sequences of joint matrix types won't be
eliminated. It results in a crash in IGC since it can't handle such case
yet. Note, it means that matrix samples compiled with -O0 also don't
work now.

So we have to (temporary?) revert this patch.

This reverts commit 6f8e45670a0b72d9b0f343c02927b709638491de.