[SYCL][ESIMD] Lower global volatile stores to vstores
According to the VC team, all stores to volatile globals need to be vstores for correctness.
Sometimes clang implicitly inserts stores, and if this happens, we need to lower to vstores.

With that, we no longer need the commit() function, so remove it and update related doc.
It never made it onto a compiler release, so we should be able to remove it with no deprecation.

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>