[SYCL][E2E] Substitute paths to lit tool binaries (#9041)
This change makes LIT explicitly substitute uses of tool binaries in
tests with their full paths. This makes test RUN lines easier to
reproduce, as it negates the effect of lit's internal environment (e.g.,
PATH) on which tools are run.

The list of tools affected are: sycl-ls llvm-spirv, llvm-link, FileCheck
and not (when used after a pipe).

This also changes the behaviour of how LIT finds sycl-ls, llvm-spirv,
and llvm-link. Whereas before these tools were found only on the PATH,
now they will be preferably sourced first from the in-tree tools
directory.