[SYCL] Throw for invalid global_work_size query (#9072)
SYCL2020:

This descriptor may only be used if the device type is
device_type::custom or if the kernel is a built-in kernel. Attempting to
use this descriptor for other devices or kernels throws an exception
with the errc::invalid error code.

The previous implementation is based on handling OpenCL exception, but
the other backends don't throw anything, so it works only on OCL. This
approch is more generic.