[SYCL][ESIMD][E2E] Fix esimd emulator timeout
This test is computationally intensive and when running with the
esimd emulator backend on a slow CPU it may hit the timeout.

Add a 512x512 case and lower the existing esimd emulator limit to 512.

This brings the testing time from 4 minutes to 1 minute.

When run on a real GPU it runs in 1 second even on Gen9 so the extra 512 case shouldn't cause any issues there.

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>