[SYCL][ESIMD] Support 64-bit offsets for block_load/block_store in stateless mode
I tried to avoid changing the parameter based on a `#define`, but we want to preserve
a warning thrown in stateful mode. My idea for that was to add a type template
for the offset type, and in stateful mode copy the offset param to a local variable of
`uint32_t` type, and that assignment would throw the shortening warning.

However it turns out we don't ever throw warnings in system headers because we pass
`-Wno-system-headers` by default, so we need the warning to occur in customer code.
I don't know any way to do that besides changing the parameter type like I do here.
`#pragma clang diagnostic push/pop` unfortunately doesn't work for this case.

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>