[SYCL][E2E] Reduce memcpy2d e2e testing (#9123)
The memcpy2d and copy2d tests in the test suite are by far the dominant
tests in execution time for the e2e tests. This is partially because of
the thorough coverage of the tests. To reduce the footprint of these
test cases, this commit removes some of the shortcut cases as they are
somewhat trivial. More extensive coverage is expected from the SYCL-CTS.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>