[SYCL][ESIMD] Fix GenXIntrinsic lowering for vectors used as scalars
In most cases, we see the following IR for vector spirv intrinsics:
```c++
%0 = load <3 x i64>, <3 x i64> addrspace(4)* @__spirv_BuiltInGlobalInvocationId, align 32
%1 = extractelement <3 x i64> %0, i64 0
// some use of %1, ex:
%2 = icmp ult i64 %0, 50
```

However, if only the first element is used, this may be optimized to:
```c++
%0 = load i64, addrspace(4)* @__spirv_BuiltInGlobalInvocationId, align 32
// some use of %1, ex:
%1 = icmp ult i64 %0, 0
```

In the latter case, if we try to insert a cast due to different types between the real intrinsic type and the global variable type,
make sure to use the load type, because the use type is arbitrary code and may not be the same as the global variable type.

This is fine for the case with ExtractElement because in that case the return will be based on the global variable type.

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>